Integrated degree centrality and test.
[SVN r51099]

+++ test/closeness_centrality.cpp
@@ -1,4 +1,4 @@
-// (C) Copyright Andrew Sutton 2007
+// (C) Copyright 2007-2009 Andrew Sutton
 //
 // Use, modification and distribution are subject to the
 // Boost Software License, Version 1.0 (See accompanying file
@@ -53,7 +53,7 @@ void test_undirected()
     typedef typename graph_traits<Graph>::vertex_descriptor Vertex;
     typedef typename graph_traits<Graph>::edge_descriptor Edge;
 
-    typedef exterior_vertex_property<Graph, float> CentralityProperty;
+    typedef exterior_vertex_property<Graph, double> CentralityProperty;
     typedef typename CentralityProperty::container_type CentralityContainer;
     typedef typename CentralityProperty::map_type CentralityMap;
 
@@ -78,11 +78,11 @@ void test_undirected()
     floyd_warshall_all_pairs_shortest_paths(g, dm, weight_map(wm));
     all_closeness_centralities(g, dm, cm);
 
-    BOOST_ASSERT(cm[v[0]] == float(1)/5);
-    BOOST_ASSERT(cm[v[1]] == float(1)/7);
-    BOOST_ASSERT(cm[v[2]] == float(1)/7);
-    BOOST_ASSERT(cm[v[3]] == float(1)/9);
-    BOOST_ASSERT(cm[v[4]] == float(1)/6);
+    BOOST_ASSERT(cm[v[0]] == double(1)/5);
+    BOOST_ASSERT(cm[v[1]] == double(1)/7);
+    BOOST_ASSERT(cm[v[2]] == double(1)/7);
+    BOOST_ASSERT(cm[v[3]] == double(1)/9);
+    BOOST_ASSERT(cm[v[4]] == double(1)/6);
 }
 
 template <typename Graph>
@@ -91,7 +91,7 @@ void test_directed()
     typedef typename graph_traits<Graph>::vertex_descriptor Vertex;
     typedef typename graph_traits<Graph>::edge_descriptor Edge;
 
-    typedef exterior_vertex_property<Graph, float> CentralityProperty;
+    typedef exterior_vertex_property<Graph, double> CentralityProperty;
     typedef typename CentralityProperty::container_type CentralityContainer;
     typedef typename CentralityProperty::map_type CentralityMap;
 
@@ -116,11 +116,11 @@ void test_directed()
     floyd_warshall_all_pairs_shortest_paths(g, dm, weight_map(wm));
     all_closeness_centralities(g, dm, cm);
 
-    BOOST_ASSERT(cm[v[0]] == float(0));
-    BOOST_ASSERT(cm[v[1]] == float(0));
-    BOOST_ASSERT(cm[v[2]] == float(0));
-    BOOST_ASSERT(cm[v[3]] == float(1)/10);
-    BOOST_ASSERT(cm[v[4]] == float(0));
+    BOOST_ASSERT(cm[v[0]] == double(0));
+    BOOST_ASSERT(cm[v[1]] == double(0));
+    BOOST_ASSERT(cm[v[2]] == double(0));
+    BOOST_ASSERT(cm[v[3]] == double(1)/10);
+    BOOST_ASSERT(cm[v[4]] == double(0));
 }
 
 int
